requirements.txt
Flask==3.1.2
Werkzeug==3.1.5
psycopg2-binary==2.9.11
python-dotenv==1.0.0
waitress>=3.0.2
supabase>=2.25.1